@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import  { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
@@ -7,13 +7,13 @@ import { Provider } from 'react-redux';
 import store from './redux/store/index';
 
 
-  ReactDOM.render(
+createRoot(document.getElementById('root')).render(
     <Provider store={store}>
       <BrowserRouter>
         
-          <App />
+        <App />
         
       </BrowserRouter>
     </Provider>,
-    document.getElementById('root')
-  );
+    
+);
