@@ -1,6 +1,6 @@
 import style from "./Landing.module.css";
 import BtnStart from "../../components/Buttons/BtnStart";
-// import imgLand from "../../Assets/cooking.png"
+import imgLand from "../../Assets/cooking.png"
 
 
 
@@ -11,11 +11,11 @@ const Landing = () => {
     return(
       <main className={style.container}>
       <div className={style.card}>
-        {/* <img src={imgLand} alt="fondo de landing" className={style.image} /> */}
-        <h1 className={style.cardTitle}>Wellcome!</h1>
+        <img src={imgLand} alt="fondo de landing" className={style.image} />
+        <h1 className={style.cardTitle}>Bienvenido!</h1>
         <div className={style.cardContent}>
-          On this site, you can find the best <span>recipes. </span>
-          In addition to the necessary <span>ingredients </span>and <span>instructions </span>to prepare them.
+          Esta pagina web fue creada por Franco Bogado, podes encontrar las mejores <span>recetas. </span>
+          Además los ingredientes <span>necesarios </span> y  <span>instrucciones</span> para prepararlos.
         </div>
         <BtnStart />
         </div>
