@@ -2,12 +2,12 @@ import Detail from "./Detail/Detail";
 import Form from "./Form/Form";
 import Home from "./Home/Home";
 import Landing from "./Landing/Landing";
-import About from "./About/About";
+
 
 export { 
     Detail, 
     Form, 
     Home, 
     Landing,
-    About
+   
 };
